fix: created self-signed cert valid for 100 years
