+++ src/__tests__/lib/localhost.crt
@@ -1,18 +1,19 @@
 -----BEGIN CERTIFICATE-----
-MIIC7TCCAdWgAwIBAgIUXlnuRfR2yE/v5t9A0ZoeVG4BvZowDQYJKoZIhvcNAQEL
-BQAwFDESMBAGA1UEAwwJbG9jYWxob3N0MB4XDTI0MDUwMTAzMTIyNloXDTI1MDUw
-MTAzMTIyNlowFDESMBAGA1UEAwwJbG9jYWxob3N0MIIBIjANBgkqhkiG9w0BAQEF
-AAOCAQ8AMIIBCgKCAQEA57iCfaiUmW74QUb0MtX5MOCOfhKJ8zw0i4Ep+nwST03w
-Z/K5Z0W7OPmJDnPsko9DPC3bOeObriTcWgYg4zNaLNSKvDbkbLtDvpUyY8rdJAP3
-6H7uTZWMoUQzdIGdSFZHa8HRO1HUZGFZT55bi7czyMPFnzSOtcaz4pCvlhLRk+QA
-lsSG+4owhfDrpyPlFtEFNyeaE2fIsJtHxupkwGOmDNeh6iKV46lD8F0SGf2pl5qF
-nbbMn6IZlpH7heQqMdPNs1ikGOuDybJISu07S72RSoClgdFzepzXFHoNWwhucdvN
-UMJXWBnP/PoeNViI2+nBMrK/1Bwuhci0t5mjTujQNQIDAQABozcwNTAUBgNVHREE
-DTALgglsb2NhbGhvc3QwHQYDVR0OBBYEFBuCDjLQbWjX7D+o9dt4nszcP3OIMA0G
-CSqGSIb3DQEBCwUAA4IBAQASgexBeY7Nz9i0qREk1RzpQDIT+jM/QAgmnH3G6Ehx
-tAYUMYLeDTmGhYhp3GJAU7/R3mbN6t5qg2d9Fa8b+JpBJRdxMY+CyjESoPvUHIE3
-lkgNGphT+8QPnh7uO5KOUVnk7Jc9MTwBntDouLHfuzJJnHPlRko3IWnwaivZYVRn
-VbnUoSMKKPzFaqqaY8uHPjvs4Gt4OGcYV8hHcjeI3fMHckmsXZclxb3pwF+x698o
-Htg5+ydbmWkTspvbMuHx/280Ow0JPSSXFnwWGWpyH7kI0EAfq75W3iGMRR6yL7Je
-ffZG7W8KARYx824nRlxbIN2rHo9VQwEBkbmoeg5nSkvi
+MIIDKTCCAhGgAwIBAgIUZRe+CwLNG0+oAye6+ZeDt3GHroAwDQYJKoZIhvcNAQEL
+BQAwPDESMBAGA1UEAwwJbG9jYWxob3N0MRkwFwYDVQQKDBBZb3VyT3JnYW5pemF0
+aW9uMQswCQYDVQQGEwJVUzAgFw0yNTA1MDUwNDUxMDJaGA8yMTI1MDQxMTA0NTEw
+MlowPDESMBAGA1UEAwwJbG9jYWxob3N0MRkwFwYDVQQKDBBZb3VyT3JnYW5pemF0
+aW9uMQswCQYDVQQGEwJVUzCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEB
+AMFeqAEOzLEVLGwFIg7JB3+Gj0TYP1x7lH7NVbAjq7pbxt7RsT6eYQsykSsGoZC5
+491o62zvO5QHjlJgYhn62t3B8fNGFTW7S4YMgyrxtavbk2E59tL+8XObZa0o0/4e
+Nu4dGQFTWm3Kv2M4bAvk1XBZsaoGlRraV6WTGmmjYzxvenPN5dfrgTemvtMl7/Ld
+h1oD+z1yTwSdSrl87lP+2P+UKu8ON2CzrH3OS1tGWiqYb+aPCh3cLsBGe0SWQ5Np
+XCs4IkW+0NQwLzkA6We//mJIhKF+1MGDSPilHaAj4mog4WdDdFb/t+tCBXaP2CGL
+GntMNikVUc7rhVE/L0jd2CsCAwEAAaMhMB8wHQYDVR0OBBYEFBcE4HHSAUAGjOX8
+UcDgNEUxID2eMA0GCSqGSIb3DQEBCwUAA4IBAQBl7VAGjTkC41vtL9Lr19z76x1S
+fDnylKjQmxVOqOvUIBmWlxJGPOEvGID9gWt+CaouSLPJGldeRp9ZKB5OZUc7TyCH
+Re7n2C0ov5DE/7xIglaqacSkMEn5JuqUaMYLyGsvJ5gkhIjo/poQmHbWWvQIG0nW
+J+H9OJvGiWgGgFye+4d3f6eB3EGka1r+ewvMDLGz/5PHOrGRmteFmMSChr0VRyEr
+UHtdzvfc5hdZnYICWns69JsBNu/kWJbF+75SuhoxdBjqjtFi1ZrfDT8WUkoQYAbc
+Mjw/5IUV5dyqoTpZCSdDziNTKo3hdZistr/tkUYX47/ssehWv2yAERzrMfGc
 -----END CERTIFICATE-----
