+++ src/__tests__/lib/localhost.key
@@ -1,28 +1,28 @@
 -----BEGIN PRIVATE KEY-----
-MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDnuIJ9qJSZbvhB
-RvQy1fkw4I5+EonzPDSLgSn6fBJPTfBn8rlnRbs4+YkOc+ySj0M8Lds545uuJNxa
-BiDjM1os1Iq8NuRsu0O+lTJjyt0kA/fofu5NlYyhRDN0gZ1IVkdrwdE7UdRkYVlP
-nluLtzPIw8WfNI61xrPikK+WEtGT5ACWxIb7ijCF8OunI+UW0QU3J5oTZ8iwm0fG
-6mTAY6YM16HqIpXjqUPwXRIZ/amXmoWdtsyfohmWkfuF5Cox082zWKQY64PJskhK
-7TtLvZFKgKWB0XN6nNcUeg1bCG5x281QwldYGc/8+h41WIjb6cEysr/UHC6FyLS3
-maNO6NA1AgMBAAECggEABihPg2JO7JLX0vg9NfqcoBFfh/cbJlUDum9Iu8i/Pc8+
-46w+BAIE2xs6aqFm+HAJI9i+Ghgzw1lkZ5oBFrh/HXDabOzyffct8isJx0AVR0c5
-t7pjv7kJHyHuUhq9oLX2OOVSF4bxo2c41dZJ7XkFTcDyB+yAHFF25z7cA5tlaU4Z
-RtIb2EDUX39A6WZKxh5S3Q33LzSvLzrT+iP3QNfTdgVAZXG8yGpOgTJdzJRzXuek
-ag+YCCcUJebqUTsRpjhlZL9Psl9UTkqtCAI+h5SiyJ3jhNiPfc+jK/mdJa4VIxcO
-F7uSxj6F9aVs8HskKlLp+2OkacNRJxCCk8SIWS7gYwKBgQD+ySm6yrZmWnnsHYpr
-tglmR+DWA5ZceBGYF/fSfUkcpFCbPChWiKvFNYyXyH+ulSncrc+YFIZq+AhhEH/1
-b3VCmNjSzSAZxeDUzQEShOC2c90GqTe05kB1jKZcLV8ow8UAbU3Bvl3n2p7wfgDx
-zoSBXUa3+LirgnN+rgSbpJqRlwKBgQDo0zUffIGr/KB3LDmIBIXqJJbFp1IQSP/9
-mQ2Pt7x8WQ+sr95uX3XmSmDn10xy6LaVuUqjDQolGpiQqV9fa48u/LOkp5eSDfa9
-2LpSHAZUJkmjAEdgAQxXl3C+Sqjue9xEgQPC1Gqo028ocBvi+5jNIRvMQzXOqWsq
-rXh+tAxOEwKBgAKeywESyKFw2MQm4Z+N9bbJknRhARDeYz/vLytJApF61yFhkwCg
-0zNqA5IkC7wd6JpOZ04nF+LNvjhRVR+PD+OpcsFoPqZFNLS7jwFDwJPfySwiXWJM
-4+jRk3xrMwlWShZhCWWusYSK6QuMIDJcb1xtOyb1NeD9rHZnD8ESBkztAoGAUKaJ
-KmUIzjjupRmOvE1EzjzqdYFmbfm7o29XQUeWZUS8p0cst/Mddi6hru3nr45pYSgU
-cLLdpsi1vLbPMNCjgtVKrWjYmmF37X+b20GV96RjsbExCFKUoer1xRU7u5DwVlKS
-qPTIFb9YLBvZrSEWZ0hmxrUpeHM7ikcZDFr2QD0CgYEAxu9Woa25BiWxpaTRlsBO
-025LK6vr+wsOlASZCJXk7Qyr5fqWwHQFOSq42wKOBL25phZiYgl/ravWoDm8QIhs
-rRYBNj9gj4l4es4KasgO3mPYSoQIVCI5ZtMUaUx56DbskJJ8RbMJstM1/I4S/g70
-OwdkJkE82BAkg9yShEuAatc=
+MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDBXqgBDsyxFSxs
+BSIOyQd/ho9E2D9ce5R+zVWwI6u6W8be0bE+nmELMpErBqGQuePdaOts7zuUB45S
+YGIZ+trdwfHzRhU1u0uGDIMq8bWr25NhOfbS/vFzm2WtKNP+HjbuHRkBU1ptyr9j
+OGwL5NVwWbGqBpUa2lelkxppo2M8b3pzzeXX64E3pr7TJe/y3YdaA/s9ck8EnUq5
+fO5T/tj/lCrvDjdgs6x9zktbRloqmG/mjwod3C7ARntElkOTaVwrOCJFvtDUMC85
+AOlnv/5iSIShftTBg0j4pR2gI+JqIOFnQ3RW/7frQgV2j9ghixp7TDYpFVHO64VR
+Py9I3dgrAgMBAAECggEABlenv+SPqVRnLHRqpWpcAiwUaFgYAaZoBcJrPqec3Vio
+bMf2hwhERJYvA1Rofje88XJTZVgEAhsae4KkXjxZtI97ztR5yDCs6QzPKAEB3E0L
+kS6tYtjeItRyJXMeX8plDo4KoOvJsM7HL/jxXJVdl/kSTyk2rn/LuidHLhL7bLVE
+IoWSWLJSFySEBMA5aAlFHGmM1awKa9FT2QqnTZgxsqCO9mAvFtZa9KyDv4ligg9F
+q0HwABkkySELsNN7RsniRwg45i4E5oeF8mUKoIlJKmiw4O6qBQng+pOLlnLNwihL
+cNQlhTrCJ0Roodmmz1vJvaF6ux0fIIEja0w+zOwaqQKBgQDxKQWudt3rthq6qkV8
+hXt423eu+eO2XLzH+f9d+FNkHEz2hGiwkra5Qyy6BZif6w1WS3VUW/5a17TwhUDQ
+SmVNRU/T3js5kiPqb2XOL23nMr53un1jioDUos0w9ddSpo3mg3O369XseW68LDmB
+aya7223bntulm1klvoVqxypxpwKBgQDNRMtYcCqiWNRwn0LE9VMEK8jNUVpJQRhf
+E/y290QIF3YtLT615SQwx6ZdwtZmsZfTBbxpe25aYjM5cFZW/m1c7mwztqt+oBkt
+GQI0ZiXkLpYp52rYgR1myh5OZAV0g2Bcs7Z7CuKvLrEVm2OdlsLVPibST7eGnMjp
+Be3VcvHN3QKBgE5C1cksYhBgZIQ3RZ2HY+5wqoJqLe+8YLc09h50u/s9A+zsQ10p
+/4pXtmgX8NPU39slWcoqoHyyi9GSLNprijxtRd6nu+iO+IDX7d3fnnbtDMqmBpML
+G4vl2VkOLvZvfpVUC3WgXygCsGD4qSDKuSM1ro0bHIS9yUxqa7mfMLL5AoGBALIW
+nIGHF0Yw4WcwlYJCAZtdzWJs0XM6ucVLlva6wlhu2oXJel3fhq2XmbxBHTiNndZk
+ZJnzj8N07VJJJ29kYEe/Rzr6ogFDL3ST/PivIvK8hGWotGdQZczXUMHcP2tdzA3B
+kXFg+VjZnqPLY/rGYRUXTiErLmPoxNp+3WBR0H41AoGAJ4yzuzwAsKtlQgBW+Ub3
+FSo70Qe3pERoRSX7n6BDdet6pUmIZB2u3t/VHfbWJxTOgmBUVx4dwrqaoGJd9AxC
+c9jPs3joKtQ+cU1CPX/I+ubUZ7sua2TO+l4g2siuA0Phbt45Bkn2OtNO1/p+sM0e
+tk4YWN/mM9rsyA53CZcLpnY=
 -----END PRIVATE KEY-----
